requirements.txt
# Google Agent Development Kit
google-genai>=1.0.0

# HTTP client for API calls
httpx>=0.27.0
requests>=2.31.0

# Free/Open-source tools
duckduckgo-search>=6.1.0  # Free web search
trafilatura>=1.12.0       # Web content extraction

# Data handling
pydantic>=2.8.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Structured output
markdown>=3.6
jinja2>=3.1.4

# Monitoring and logging
loguru>=0.7.2

# File handling
aiofiles>=24.1.0

# Web interface
flask>=3.0.0
flask-cors>=4.0.0  # For CORS support if needed

# Testing
pytest>=8.3.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0

# Utilities
python-dateutil>=2.9.0
tenacity>=8.5.0  # Retry logic

